add minimap2 support

+++ jTAD80.bash
@@ -50,11 +50,11 @@ done
 
 dfiles="${map_out}/*.fasta"
 
-if ! command -v samtools &> /dev/null
+if ! command -v samtools && ! command -v bedtools &> /dev/null
 then
     $(ls $dfiles | parallel -j $processors "./dependencies/samtools_linux sort -n -O bam -o {.}.byread {.}.sorted.bam")
     byreads_bam="${map_out}/*.byread"
-    $(ls $byreads_bam | parallel -j $processors "./dependencies/filterBam_linux --in {} --out {.}.filtered --minCover $coverage --minId $identity")
+    $(ls $byreads_bam | parallel -j $processors "./dependencies/filterBam_linux --best --in {} --out {.}.filtered --minCover $coverage --minId $identity")
     $(rm $byreads_bam)
     filtered_bam="${map_out}/*.filtered"
     $(ls $filtered_bam | parallel -j $processors "./dependencies/samtools_linux sort -O bam -o {.}.final {}")
@@ -64,13 +64,14 @@ then
 else
     $(ls $dfiles | parallel -j $processors "samtools sort -n -O bam -o {.}.byread {.}.sorted.bam")
     byreads_bam="${map_out}/*.byread"
-    $(ls $byreads_bam | parallel -j $processors "./dependencies/filterBam_linux --in {} --out {.}.filtered --minCover $coverage --minId $identity")
+    #$(ls $byreads_bam | parallel -j $processors "./dependencies/filterBam_linux --best --in {} --out {.}.filtered --minCover $coverage --minId $identity")
+    $(ls $byreads_bam | parallel -j $processors "./dependencies/filterBam_linux --uniqThresh 0.90 --in {} --out {.}.filtered --minCover $coverage --minId $identity")
     $(rm $byreads_bam)
     filtered_bam="${map_out}/*.filtered"
     $(ls $filtered_bam | parallel -j $processors "samtools sort -O bam -o {.}.final {}")
     $(rm $filtered_bam)
     final_bam="${map_out}/*.final"
-    $(ls $final_bam | parallel -j $processors "./dependencies/bedtools_linux genomecov -ibam {} -bga > {.}.depth")
+    $(ls $final_bam | parallel -j $processors "bedtools genomecov -ibam {} -bga > {.}.depth")
 fi
 
 for F in $final_bam; do
